+++ src/devtools/PLandDevTools.cc
@@ -1,3 +1,5 @@
+#include "fmt/format.h"
+#include <filesystem>
 #ifdef LD_DEVTOOL
 #include "PLandDevTools.h"
 #include "devtools/impl/DataMenu.h"
@@ -115,11 +117,24 @@ void PLandDevTools::_updateScale() {
     if (prevScale != xScale) {
         prevScale   = xScale;
         ImGuiIO& io = ImGui::GetIO();
-        // auto     fontPath = "C:/Windows/Fonts/msyh.ttc";
-        // auto     fontPath = "D:/Datas/Desktop/imgui-1.91.6-docking/misc/fonts/Cousine-Regular.ttf";
-        auto fontPath = "D:/Datas/Desktop/MapleMonoNormal-CN-Regular.ttf";
+
+        auto fontPath = my_mod::MyMod::getInstance().getSelf().getDataDir() / "fonts" / "font.ttf";
+        if (!std::filesystem::exists(fontPath)) {
+            this->warning(fmt::format(
+                "由于字体文件 ( {} ) 不存在\n这可能导致部分模块字体显示异常\n\n建议下载 maple-font 字体的 "
+                "Normal-Ligature CN 版本\n将 \"MapleMonoNormal-CN-Regular.ttf\" 重命名为 font.ttf 放置在上述路径下",
+                fontPath.string()
+            ));
+            fontPath = "C:/Windows/Fonts/msyh.ttc";
+        }
+
         io.Fonts->Clear();
-        io.Fonts->AddFontFromFileTTF(fontPath, std::round(15 * xScale), nullptr, io.Fonts->GetGlyphRangesChineseFull());
+        io.Fonts->AddFontFromFileTTF(
+            fontPath.string().c_str(),
+            std::round(15 * xScale),
+            nullptr,
+            io.Fonts->GetGlyphRangesChineseFull()
+        );
         io.Fonts->Build();
         ImGui_ImplOpenGL3_DestroyFontsTexture();
         ImGui_ImplOpenGL3_CreateFontsTexture();
@@ -146,6 +161,23 @@ void PLandDevTools::_renderMenuBar() {
     }
 }
 
+void PLandDevTools::warning(std::string msg) { this->warnings_.emplace(std::move(msg)); }
+
+void PLandDevTools::_renderWarnings() {
+    if (!this->warnings_.empty()) {
+        if (!ImGui::Begin("Warning")) {
+            ImGui::End();
+            return;
+        }
+        ImGui::PushStyleColor(ImGuiCol_Text, IM_COL32(255, 40, 40, 255));
+        for (auto& warning : this->warnings_) {
+            ImGui::TextUnformatted(warning.data());
+        }
+        ImGui::PopStyleColor();
+        ImGui::End();
+    }
+}
+
 void PLandDevTools::render() {
     if (!this->isInited_) {
         this->_init();
@@ -167,7 +199,10 @@ void PLandDevTools::render() {
         ImGui_ImplGlfw_NewFrame();
         ImGui::NewFrame();
 
-        _renderMenuBar(); // 渲染菜单栏
+        _renderMenuBar();  // 渲染菜单栏
+        _renderWarnings(); // 渲染警告
+
+        // 更新组件状态
         for (auto& comp : this->menus_) {
             comp->tick();
         }