+++ src/devtools/PLandDevTools.h
@@ -1,3 +1,5 @@
+#include <unordered_map>
+#include <unordered_set>
 #ifdef LD_DEVTOOL
 #pragma once
 #include "imgui.h"
@@ -26,17 +28,21 @@ public:
 
     void hide();
 
+    void warning(std::string msg);
+
 private:
     void render();
-    void _init();          // 初始化 ImGui & Opengl & GLFW
-    void _renderMenuBar(); // 渲染菜单栏
-    void _updateScale();   // 更新缩放比例
+    void _init();           // 初始化 ImGui & Opengl & GLFW
+    void _renderMenuBar();  // 渲染菜单栏
+    void _updateScale();    // 更新缩放比例
+    void _renderWarnings(); // 渲染警告信息
 
     bool                                        isInited_{false};            // 是否初始化 _init
     GLFWwindow*                                 window_{nullptr};            // 窗口
     std::thread                                 renderThread_;               // 渲染线程
     std::atomic_bool                            renderThreadRunning_{false}; // 渲染线程是否运行
     std::vector<std::unique_ptr<MenuComponent>> menus_;                      // 菜单栏
+    std::unordered_set<std::string>             warnings_;                   // 警告信息
 };
 
 inline std::unique_ptr<PLandDevTools> g_devtools = nullptr;