+++ CHANGELOG.md
@@ -24,6 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - 优化 `PLand::getLandAt` 查询
 - 重构 `Calculate` 为 `PriceCalculate`
 - 适配事件库 v0.4.0 #24
+- 重构 `PLand DevTools` 开发者工具
 - 重构部分代码
 
 ### Fixed